Move test helpers around
