+++ tests/helpers/defer.js
@@ -1,0 +1,13 @@
+const { assign } = Object;
+
+export default function defer() {
+  let deferred = {};
+
+  let promise = new Promise((resolve, reject) => {
+    assign(deferred, { resolve, reject });
+  });
+
+  return assign(deferred, {
+    then: promise.then.bind(promise)
+  });
+}
